Move the common methods to the DeviceWrapper trait
This avoids duplicate code
