+++ src/lib.rs
@@ -54,11 +54,9 @@ pub mod uinput;
 pub mod util;
 
 use bitflags::bitflags;
-use libc::{c_uint, c_void, suseconds_t, time_t};
-use std::any::Any;
+use libc::{c_uint, suseconds_t, time_t};
 use std::convert::{TryFrom, TryInto};
 use std::time::{Duration, SystemTime, SystemTimeError, UNIX_EPOCH};
-use std::{io, ptr};
 
 use enums::*;
 use util::*;
@@ -67,6 +65,8 @@ use evdev_sys as raw;
 
 #[doc(inline)]
 pub use device::Device;
+#[doc(inline)]
+pub use device::DeviceWrapper;
 #[doc(inline)]
 pub use device::UninitDevice;
 #[doc(inline)]
@@ -269,232 +269,5 @@ impl InputEvent {
         let (ev_type, ev_code) = event_code_to_int(code);
 
         unsafe { raw::libevdev_event_is_code(&self.as_raw(), ev_type, ev_code) == 1 }
-    }
-}
-
-/// Abstraction over structs which contain an inner `*mut libevdev`
-pub trait LibevdevWrapper {
-    fn raw(&self) -> *mut raw::libevdev;
-
-    /// Forcibly enable an EventType/InputProp on this device, even if the underlying
-    /// device does not support it. While this cannot make the device actually
-    /// report such events, it will now return true for has().
-    ///
-    /// This is a local modification only affecting only this representation of
-    /// this device.
-    fn enable(&self, blob: &dyn Any) -> io::Result<()> {
-        if let Some(ev_type) = blob.downcast_ref::<EventType>() {
-            self.enable_event_type(ev_type)
-        } else if let Some(ev_code) = blob.downcast_ref::<EventCode>() {
-            self.enable_event_code(ev_code, None)
-        } else if let Some(prop) = blob.downcast_ref::<InputProp>() {
-            self.enable_property(prop)
-        } else {
-            Err(io::Error::from_raw_os_error(-1))
-        }
-    }
-
-    /// Enables this property, a call to `set_file` will overwrite any previously set values
-    ///
-    /// Note: Please use the `enable` function instead. This function is only
-    /// available for the sake of maintaining compatibility with libevdev.
-    fn enable_property(&self, prop: &InputProp) -> io::Result<()> {
-        let result =
-            unsafe { raw::libevdev_enable_property(self.raw(), *prop as c_uint) as i32 };
-
-        match result {
-            0 => Ok(()),
-            error => Err(io::Error::from_raw_os_error(-error)),
-        }
-    }
-
-    /// Forcibly enable an event type on this device, even if the underlying
-    /// device does not support it. While this cannot make the device actually
-    /// report such events, it will now return true for libevdev_has_event_type().
-    ///
-    /// This is a local modification only affecting only this representation of
-    /// this device.
-    ///
-    /// Note: Please use the `enable` function instead. This function is only
-    /// available for the sake of maintaining compatibility with libevdev.
-    fn enable_event_type(&self, ev_type: &EventType) -> io::Result<()> {
-        let result =
-            unsafe { raw::libevdev_enable_event_type(self.raw(), *ev_type as c_uint) };
-
-        match result {
-            0 => Ok(()),
-            error => Err(io::Error::from_raw_os_error(-error)),
-        }
-    }
-
-    /// Forcibly enable an event type on this device, even if the underlying
-    /// device does not support it. While this cannot make the device actually
-    /// report such events, it will now return true for libevdev_has_event_code().
-    ///
-    /// The last argument depends on the type and code:
-    /// If type is EV_ABS, data must be a pointer to a struct input_absinfo
-    /// containing the data for this axis.
-    /// If type is EV_REP, data must be a pointer to a int containing the data
-    /// for this axis.
-    /// For all other types, the argument must be NULL.
-    ///
-    /// Note: Please use the `enable` function instead. This function is only
-    /// available for the sake of maintaining compatibility with libevdev.
-    fn enable_event_code(
-        &self,
-        ev_code: &EventCode,
-        blob: Option<&dyn Any>,
-    ) -> io::Result<()> {
-        let (ev_type, ev_code) = event_code_to_int(ev_code);
-
-        let data = blob
-            .map(|data| {
-                data.downcast_ref::<AbsInfo>()
-                    .map(|absinfo| &absinfo.as_raw() as *const _ as *const c_void)
-                    .unwrap_or_else(|| data as *const _ as *const c_void)
-            })
-            .unwrap_or_else(|| ptr::null() as *const _ as *const c_void);
-
-        let result = unsafe {
-            raw::libevdev_enable_event_code(
-                self.raw(),
-                ev_type as c_uint,
-                ev_code as c_uint,
-                data,
-            )
-        };
-
-        match result {
-            0 => Ok(()),
-            error => Err(io::Error::from_raw_os_error(-error)),
-        }
-    }
-
-    /// Forcibly disable an EventType/EventCode on this device, even if the
-    /// underlying device provides it. This effectively mutes the respective set of
-    /// events. has() will return false for this EventType/EventCode
-    ///
-    /// In most cases, a caller likely only wants to disable a single code, not
-    /// the whole type.
-    ///
-    /// Disabling EV_SYN will not work. In Peter's Words "Don't shoot yourself
-    /// in the foot. It hurts".
-    ///
-    /// This is a local modification only affecting only this representation of
-    /// this device.
-    fn disable(&self, blob: &dyn Any) -> io::Result<()> {
-        if let Some(ev_type) = blob.downcast_ref::<EventType>() {
-            self.disable_event_type(ev_type)
-        } else if let Some(ev_code) = blob.downcast_ref::<EventCode>() {
-            self.disable_event_code(ev_code)
-        } else {
-            Err(io::Error::from_raw_os_error(-1))
-        }
-    }
-
-    /// Forcibly disable an event type on this device, even if the underlying
-    /// device provides it. This effectively mutes the respective set of
-    /// events. libevdev will filter any events matching this type and none will
-    /// reach the caller. libevdev_has_event_type() will return false for this
-    /// type.
-    ///
-    /// In most cases, a caller likely only wants to disable a single code, not
-    /// the whole type. Use `disable_event_code` for that.
-    ///
-    /// Disabling EV_SYN will not work. In Peter's Words "Don't shoot yourself
-    /// in the foot. It hurts".
-    ///
-    /// This is a local modification only affecting only this representation of
-    /// this device.
-    ///
-    /// Note: Please use the `disable` function instead. This function is only
-    /// available for the sake of maintaining compatibility with libevdev.
-    fn disable_event_type(&self, ev_type: &EventType) -> io::Result<()> {
-        let result =
-            unsafe { raw::libevdev_disable_event_type(self.raw(), *ev_type as c_uint) };
-
-        match result {
-            0 => Ok(()),
-            error => Err(io::Error::from_raw_os_error(-error)),
-        }
-    }
-    /// Forcibly disable an event code on this device, even if the underlying
-    /// device provides it. This effectively mutes the respective set of
-    /// events. libevdev will filter any events matching this type and code and
-    /// none will reach the caller. `has_event_code` will return false for
-    /// this code.
-    ///
-    /// Disabling all event codes for a given type will not disable the event
-    /// type. Use `disable_event_type` for that.
-    ///
-    /// This is a local modification only affecting only this representation of
-    /// this device.
-    ///
-    /// Disabling codes of type EV_SYN will not work. Don't shoot yourself in the
-    /// foot. It hurts.
-    ///
-    /// Note: Please use the `disable` function instead. This function is only
-    /// available for the sake of maintaining compatibility with libevdev.
-    fn disable_event_code(&self, code: &EventCode) -> io::Result<()> {
-        let (ev_type, ev_code) = event_code_to_int(code);
-        let result =
-            unsafe { raw::libevdev_disable_event_code(self.raw(), ev_type, ev_code) };
-
-        match result {
-            0 => Ok(()),
-            error => Err(io::Error::from_raw_os_error(-error)),
-        }
-    }
-
-    /// Returns `true` if device support the InputProp/EventType/EventCode and false otherwise
-    fn has(&self, blob: &dyn Any) -> bool {
-        if let Some(ev_type) = blob.downcast_ref::<EventType>() {
-            self.has_event_type(ev_type)
-        } else if let Some(ev_code) = blob.downcast_ref::<EventCode>() {
-            self.has_event_code(ev_code)
-        } else if let Some(prop) = blob.downcast_ref::<InputProp>() {
-            self.has_property(prop)
-        } else {
-            false
-        }
-    }
-
-    /// Returns `true` if device support the property and false otherwise
-    ///
-    /// Note: Please use the `has` function instead. This function is only
-    /// available for the sake of maintaining compatibility with libevdev.
-    fn has_property(&self, prop: &InputProp) -> bool {
-        unsafe { raw::libevdev_has_property(self.raw(), *prop as c_uint) != 0 }
-    }
-
-    /// Returns `true` is the device support this event type and `false` otherwise
-    ///
-    /// Note: Please use the `has` function instead. This function is only
-    /// available for the sake of maintaining compatibility with libevdev.
-    fn has_event_type(&self, ev_type: &EventType) -> bool {
-        unsafe { raw::libevdev_has_event_type(self.raw(), *ev_type as c_uint) != 0 }
-    }
-
-    /// Return `true` is the device support this event type and code and `false` otherwise
-    ///
-    /// Note: Please use the `has` function instead. This function is only
-    /// available for the sake of maintaining compatibility with libevdev.
-    fn has_event_code(&self, code: &EventCode) -> bool {
-        unsafe {
-            let (ev_type, ev_code) = event_code_to_int(code);
-            raw::libevdev_has_event_code(self.raw(), ev_type, ev_code) != 0
-        }
-    }
-}
-
-impl LibevdevWrapper for UninitDevice {
-    fn raw(&self) -> *mut raw::libevdev {
-        self.raw
-    }
-}
-
-impl LibevdevWrapper for Device {
-    fn raw(&self) -> *mut raw::libevdev {
-        self.raw
     }
 }
