+++ src/macros.rs
@@ -2,7 +2,7 @@ macro_rules! string_getter {
     ( $( #[$doc:meta], $func_name:ident, $c_func: ident ),* ) => {
         $(
             #[$doc]
-            pub fn $func_name (&self) -> Option<&str> {
+            fn $func_name (&self) -> Option<&str> {
                 unsafe {
                     ptr_to_str(raw::$c_func(self.raw()))
                 }
@@ -14,7 +14,7 @@ macro_rules! string_getter {
 macro_rules! string_setter {
     ( $( $func_name:ident, $c_func: ident ),* ) => {
         $(
-            pub fn $func_name (&self, field: &str) {
+            fn $func_name (&self, field: &str) {
                 let field = CString::new(field).unwrap();
                 unsafe {
                     raw::$c_func(self.raw(), field.as_ptr())
@@ -27,7 +27,7 @@ macro_rules! string_setter {
 macro_rules! product_getter {
     ( $( $func_name:ident, $c_func: ident ),* ) => {
         $(
-            pub fn $func_name (&self) -> u16 {
+            fn $func_name (&self) -> u16 {
                 unsafe {
                     raw::$c_func(self.raw()) as u16
                 }
@@ -39,7 +39,7 @@ macro_rules! product_getter {
 macro_rules! product_setter {
     ( $( $func_name:ident, $c_func: ident ),* ) => {
         $(
-            pub fn $func_name (&self, field: u16) {
+            fn $func_name (&self, field: u16) {
                 unsafe {
                     raw::$c_func(self.raw(), field as c_int);
                 }
@@ -51,7 +51,7 @@ macro_rules! product_setter {
 macro_rules! abs_getter {
     ( $( $func_name:ident, $c_func: ident ),* ) => {
         $(
-            pub fn $func_name (&self,
+            fn $func_name (&self,
                                code: u32) -> std::io::Result<i32> {
                 let result = unsafe {
                     raw::$c_func(self.raw(), code as c_uint) as i32
@@ -69,7 +69,7 @@ macro_rules! abs_getter {
 macro_rules! abs_setter {
     ( $( $func_name:ident, $c_func: ident ),* ) => {
         $(
-            pub fn $func_name (&self,
+            fn $func_name (&self,
                                code: u32,
                                val: i32) {
                 unsafe {
